Minor Cygwin bugfixes
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3542 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_dictionary.c
@@ -169,12 +169,13 @@ mx_dictionary_show_dictionary( MX_DICTIONARY *dictionary )
 			    array_index = MX_ARRAY_OFFSET_NUM_DIMENSIONS - 1;
 
 			    fprintf( stderr, ", array = (%ld",
-				header[array_index] );
+				(long) header[array_index] );
 
 			    for ( j = 1; j < num_dimensions; j++ ) {
 				array_index--;
 
-				fprintf( stderr, ",%ld", header[array_index] );
+				fprintf( stderr, ",%ld",
+					(long) header[array_index] );
 			    }
 			    fprintf( stderr, ")" );
 			}
